Fix uncompatible package check and modify the base composer class
